requirements.txt
openai>=0.28.0
langchain>=0.0.283
pinecone-client>=2.2.2
python-dotenv>=1.0.0
tiktoken>=0.4.0
ipython>=8.15.0
notebook>=7.0.3
jupyterlab>=4.0.5
pypdf>=3.16.0
wikipedia>=1.4.0
